Initial work for 7.0 (#80)
* setup branch for endwalker updates

* resharper updates, rpr, sge, and file cache issues

* updates and other QOL items

* Changes for Endwalker 6.3 Support

Also switched to locally hosted struct and sig json files.

* More updates

* Added some checks for null references and added keybind charges.

* Initial 7.0 work

---------

+++ Sharlayan/Models/Structures/JobResources.Bard.cs
@@ -1,6 +1,6 @@
 ﻿// --------------------------------------------------------------------------------------------------------------------
 // <copyright file="JobResources.Bard.cs" company="SyndicatedLife">
-//   Copyright© 2007 - 2021 Ryan Wilson <syndicated.life@gmail.com> (https://syndicated.life/)
+//   Copyright© 2007 - 2022 Ryan Wilson <syndicated.life@gmail.com> (https://syndicated.life/)
 //   Licensed under the MIT license. See LICENSE.md in the solution root for full license information.
 // </copyright>
 // <summary>
@@ -11,7 +11,7 @@
 namespace Sharlayan.Models.Structures {
     public partial class JobResources {
         public sealed class BardResources {
-            public int ActiveSong { get; set; }
+            public byte ActiveSong { get; set; }
             public int Repertoire { get; set; }
             public int SoulVoice { get; set; }
             public int Timer { get; set; }
